[#28] Rearrange Models folder
