+++ MKVStudio/Models/Objects/ProcessResult/ProcessResultNames.cs
@@ -1,0 +1,13 @@
+﻿namespace MKVStudio.Models;
+
+public enum ProcessResultNames
+{
+    LoudnormFirst,
+    LoudnormSecondStereo,
+    LoudnormSecond6Channels,
+    MKVInfo,
+    MKVExtractAttachments,
+    MKVMergeJ,
+    MKVMergeLangList,
+    MKVMergeSupportedFileTypes
+}
